Separate out unmarshal functions into their own file
