+++ .vscode/settings.json
@@ -74,6 +74,7 @@
             "Tracef",
             "unmarshal",
             "unmarshalled",
+            "unmarshallers",
             "unmarshalling",
             "unmarshals",
             "usecases",
