+++ license.json
@@ -8,7 +8,8 @@
                   "annotationRefs": ["APPROVED"],
                   "aliases":["Free Public License 1.0.0"],
                   "notes": [""],
-                  "urls": ["https://spdx.org/licenses/0BSD"]
+                  "urls": ["https://spdx.org/licenses/0BSD",
+                           "https://opensource.org/licenses/0BSD"]
             },
             {
                   "id": "",
